+++ bittytax/price/pricedata.py
@@ -7,7 +7,8 @@ from colorama import Fore
 
 from ..version import __version__
 from ..config import config
-from .datasource import ExchangeRatesAPI, RatesAPI, CoinDesk, CryptoCompare, CoinGecko, CoinPaprika
+from .datasource import DataSourceBase
+from .exceptions import UnexpectedDataSourceError
 
 class PriceData(object):
     def __init__(self):
@@ -20,11 +21,9 @@ class PriceData(object):
                                     config.data_source_crypto) | \
                                 {x for v in config.data_source_select.values() for x in v}
 
-        for data_source in data_sources_required:
-            try:
-                self.data_sources[data_source] = globals()[data_source]()
-            except KeyError:
-                raise ValueError("data source: %s not recognised" % [data_source])
+        for data_source in DataSourceBase.__subclasses__():
+            if data_source.__name__.upper() in [ds.upper() for ds in data_sources_required]:
+                self.data_sources[data_source.__name__.upper()] = data_source()
 
     @staticmethod
     def data_source_priority(asset):
@@ -36,41 +35,41 @@ class PriceData(object):
             return config.data_source_crypto
 
     def get_latest_ds(self, data_source, asset, quote):
-        if data_source in self.data_sources:
-            if asset in self.data_sources[data_source].assets:
-                return self.data_sources[data_source].get_latest(asset, quote), \
-                       self.data_sources[data_source].assets[asset]
+        if data_source.upper() in self.data_sources:
+            if asset in self.data_sources[data_source.upper()].assets:
+                return self.data_sources[data_source.upper()].get_latest(asset, quote), \
+                       self.data_sources[data_source.upper()].assets[asset]
 
             return None, None
         else:
-            raise ValueError("data source: %s not recognised" % [data_source])
+            raise UnexpectedDataSourceError(data_source)
 
     def get_historical_ds(self, data_source, asset, quote, timestamp):
-        if data_source in self.data_sources:
-            if asset in self.data_sources[data_source].assets:
+        if data_source.upper() in self.data_sources:
+            if asset in self.data_sources[data_source.upper()].assets:
                 date = timestamp.strftime('%Y-%m-%d')
                 pair = asset + '/' + quote
 
                 if not config.args.nocache:
                     # check cache first
-                    if pair in self.data_sources[data_source].prices and \
-                       date in self.data_sources[data_source].prices[pair]:
-                        return self.data_sources[data_source].prices[pair][date]['price'], \
-                               self.data_sources[data_source].assets[asset], \
-                               self.data_sources[data_source].prices[pair][date]['url']
+                    if pair in self.data_sources[data_source.upper()].prices and \
+                       date in self.data_sources[data_source.upper()].prices[pair]:
+                        return self.data_sources[data_source.upper()].prices[pair][date]['price'], \
+                              self.data_sources[data_source.upper()].assets[asset], \
+                              self.data_sources[data_source.upper()].prices[pair][date]['url']
 
-                self.data_sources[data_source].get_historical(asset, quote, timestamp)
-                if pair in self.data_sources[data_source].prices and \
-                   date in self.data_sources[data_source].prices[pair]:
-                    return self.data_sources[data_source].prices[pair][date]['price'], \
-                           self.data_sources[data_source].assets[asset], \
-                           self.data_sources[data_source].prices[pair][date]['url']
+                self.data_sources[data_source.upper()].get_historical(asset, quote, timestamp)
+                if pair in self.data_sources[data_source.upper()].prices and \
+                   date in self.data_sources[data_source.upper()].prices[pair]:
+                    return self.data_sources[data_source.upper()].prices[pair][date]['price'], \
+                           self.data_sources[data_source.upper()].assets[asset], \
+                           self.data_sources[data_source.upper()].prices[pair][date]['url']
                 else:
-                    return None, self.data_sources[data_source].assets[asset], None
+                    return None, self.data_sources[data_source.upper()].assets[asset], None
             else:
                 return None, None, None
         else:
-            raise ValueError("data source: %s not recognised" % [data_source])
+            raise UnexpectedDataSourceError(data_source)
 
     def get_latest(self, asset, quote):
         price = name = data_source = None
@@ -83,11 +82,11 @@ class PriceData(object):
                         asset,
                         '{:0,f}'.format(price.normalize()),
                         quote,
-                        data_source,
+                        self.data_sources[data_source.upper()].name(),
                         name))
                 break
 
-        return price, name, data_source
+        return price, name, self.data_sources[data_source.upper()].name()
 
     def get_historical(self, asset, quote, timestamp):
         price = name = data_source = url = None
@@ -101,8 +100,8 @@ class PriceData(object):
                         asset,
                         '{:0,f}'.format(price.normalize()),
                         quote,
-                        data_source,
+                        self.data_sources[data_source.upper()].name(),
                         name))
                 break
 
-        return price, name, data_source, url
+        return price, name, self.data_sources[data_source.upper()].name(), url